Refactor portal specific colors out of react-shared
