+++ packages/react-shared/CHANGELOG.json
@@ -1,5 +1,13 @@
 [
     {
+        "version": "0.2.0",
+        "changes": [
+            {
+                "note": "Removed portal specific colors"
+            }
+        ]
+    },
+    {
         "timestamp": 1525477860,
         "version": "0.1.6",
         "changes": [
