+++ packages/website/ts/components/portal/back_button.tsx
@@ -1,6 +1,8 @@
-import { colors, Styles } from '@0xproject/react-shared';
+import { Styles } from '@0xproject/react-shared';
 import * as React from 'react';
 import { Link } from 'react-router-dom';
+
+import { colors } from 'ts/utils/colors';
 
 export interface BackButtonProps {
     to: string;
