+++ packages/website/ts/components/wallet/wallet.tsx
@@ -1,6 +1,5 @@
 import { ZeroEx } from '0x.js';
 import {
-    colors,
     constants as sharedConstants,
     EtherscanLinkSuffixes,
     Styles,
@@ -42,6 +41,7 @@ import {
     WebsitePaths,
 } from 'ts/types';
 import { backendClient } from 'ts/utils/backend_client';
+import { colors } from 'ts/utils/colors';
 import { constants } from 'ts/utils/constants';
 import { utils } from 'ts/utils/utils';
 import { styles as walletItemStyles } from 'ts/utils/wallet_item_styles';
