+++ packages/website/ts/utils/wallet_item_styles.ts
@@ -1,4 +1,6 @@
-import { colors, Styles } from '@0xproject/react-shared';
+import { Styles } from '@0xproject/react-shared';
+
+import { colors } from 'ts/utils/colors';
 
 export const styles: Styles = {
     focusedItem: {
